requirements.txt

GDAL==3.3.2   
geopandas==0.11.0 
keras==2.10.0
matplotlib==3.6.0
patchify==0.2.3  
PyYAML==5.3.1     
rasterio==1.3.2
requests==2.26.0
scipy==1.9.1
sklearn==0.0
tensorflow==2.10.0
tifffile==2022.8.12   